feat: alterada interacao de arquivos de environment no entrypoint.sh

+++ .docker/app/entrypoint.sh
@@ -2,8 +2,15 @@
 
 #On error no such file entrypoint.sh, execute in terminal - dos2unix .docker\entrypoint.sh
 chown -R www-data:www-data .
-cp .env.example .env
-cp .env.testing.example .env.testing
+
+if [ ! -f ".env" ]; then
+    cp .env.example .env
+fi
+
+if [ ! -f ".env.testing" ]; then
+    cp .env.testing.example .env.testing
+fi
+
 composer install
 php artisan key:generate
 php artisan migrate
